Finished Events Page & Updated JS
finished the events page with sample content and images. created place for information about the events, what to expect, and resources such as parking information that may be helpful to attendees. also, updated JS to add a quality of life fix. the fix prevents fade and unfading the active page onclick.